WEA-35 Move temperature conversion to presenter
