+++ app/src/main/java/com/ayush/weatherapp/home/HomePresenterImpl.java
@@ -10,14 +10,21 @@ import android.support.v4.app.ActivityCompat;
 import com.ayush.weatherapp.R;
 import com.ayush.weatherapp.constants.Temperature;
 import com.ayush.weatherapp.constants.TemperatureUnit;
+import com.ayush.weatherapp.entities.forecast.CurrentForecastEntity;
+import com.ayush.weatherapp.entities.forecast.DailyDataEntity;
 import com.ayush.weatherapp.entities.forecast.ForecastEntity;
+import com.ayush.weatherapp.entities.forecast.HourlyDataEntity;
 import com.ayush.weatherapp.entities.geocoding.GeolocationEntity;
 import com.ayush.weatherapp.mvp.BasePresenterImpl;
+import com.ayush.weatherapp.realm.RealmUtils;
+import com.ayush.weatherapp.realm.model.forecast.Forecast;
+import com.ayush.weatherapp.repository.forecast.ForecastRepository;
+import com.ayush.weatherapp.repository.forecast.ForecastRepositoryImpl;
 import com.ayush.weatherapp.repository.geocoding.GeocodingRepository;
 import com.ayush.weatherapp.repository.geocoding.GeocodingRepositoryImpl;
 import com.ayush.weatherapp.repository.preferences.PreferenceRepository;
 import com.ayush.weatherapp.repository.preferences.PreferenceRepositoryImpl;
-import com.ayush.weatherapp.repository.forecast.ForecastRepositoryImpl;
+import com.ayush.weatherapp.utils.UnitConversionUtils;
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationCallback;
 import com.google.android.gms.location.LocationRequest;
@@ -27,20 +34,22 @@ import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.disposables.Disposable;
 import io.reactivex.observers.DisposableObserver;
 import io.reactivex.schedulers.Schedulers;
+import io.realm.Realm;
+import java.util.List;
 import timber.log.Timber;
 
 public class HomePresenterImpl extends BasePresenterImpl<HomeContract.View>
     implements HomeContract.Presenter {
   private static final int LOCATION_REQ_INTERVAL = 10000;
   private static final int FASTEST_LOCATION_REQ_INTERVAL = 5000;
-
+  @Temperature private static int defaultTemperatureUnit = TemperatureUnit.FAHRENHEIT;
   private FusedLocationProviderClient fusedLocationProviderClient;
   private LocationRequest locationRequest;
   private LocationCallback locationCallback;
 
   private PreferenceRepository preferenceRepository;
   private ForecastEntity forecast;
-  private ForecastRepositoryImpl weatherRepositoryImpl;
+  private ForecastRepository forecastRepository;
   private GeocodingRepository geocodingRepository;
 
   // TODO dagger
@@ -52,7 +61,7 @@ public class HomePresenterImpl extends BasePresenterImpl<HomeContract.View>
     super.attachView(view);
     fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(getContext());
     preferenceRepository.onPreferenceChangeListener(newTemperature -> setForecastView());
-    weatherRepositoryImpl = new ForecastRepositoryImpl();
+    forecastRepository = new ForecastRepositoryImpl();
     geocodingRepository = new GeocodingRepositoryImpl();
   }
 
@@ -198,12 +207,15 @@ public class HomePresenterImpl extends BasePresenterImpl<HomeContract.View>
   }
 
   private void fetchWeatherForecast(double lat, double lng, boolean isCurrentLocation) {
-    Disposable disposable = weatherRepositoryImpl.getForecast(lat, lng, isCurrentLocation)
+    Disposable disposable = forecastRepository.getForecast(lat, lng, isCurrentLocation)
         .subscribeOn(Schedulers.io())
         .observeOn(AndroidSchedulers.mainThread())
         .doOnSubscribe(d -> getView().showProgressBar(""))
         .subscribeWith(new DisposableObserver<ForecastEntity>() {
           @Override public void onNext(ForecastEntity forecast) {
+            //initialize value
+            defaultTemperatureUnit = TemperatureUnit.FAHRENHEIT;
+
             setForecast(forecast, lat, lng);
 
             //save to provide coordinates during refresh
@@ -242,7 +254,7 @@ public class HomePresenterImpl extends BasePresenterImpl<HomeContract.View>
       return;
     }
 
-    weatherRepositoryImpl.checkTemperatureUnit(forecast);
+    checkTemperatureUnit(forecast);
     getView().setDailyForeCast(forecast.getDailyForecastEntity().getDailyDataEntityList());
     getView().setHourlyForeCast(forecast.getHourlyForecastEntity().getHourlyDataEntityList());
     getView().setCurrentForecast(forecast.getCurrentForecastEntity());
@@ -268,5 +280,62 @@ public class HomePresenterImpl extends BasePresenterImpl<HomeContract.View>
       return false;
     }
     return true;
+  }
+
+  public void checkTemperatureUnit(ForecastEntity forecast) {
+    //TODO logic is not elegant
+    if (preferenceRepository.getTemperatureUnit() != defaultTemperatureUnit) {
+      convertCurrentTemperature(forecast.getCurrentForecastEntity());
+      convertDailyData(forecast.getDailyForecastEntity().getDailyDataEntityList());
+      convertHourlyData(forecast.getHourlyForecastEntity().getHourlyDataEntityList());
+
+      defaultTemperatureUnit = preferenceRepository.getTemperatureUnit();
+    }
+  }
+
+  private void convertDailyData(List<DailyDataEntity> dailyDatas) {
+    if (preferenceRepository.getTemperatureUnit() == TemperatureUnit.CELSIUS) {
+      for (DailyDataEntity dailyData : dailyDatas) {
+        dailyData.setWindSpeed(UnitConversionUtils.mphToKmph(dailyData.getWindSpeed()));
+        dailyData.setTemperatureHigh(
+            (int) UnitConversionUtils.fahrenheitToCelsius(dailyData.getTemperatureHigh()));
+        dailyData.setTemperatureLow(
+            (int) UnitConversionUtils.fahrenheitToCelsius(dailyData.getTemperatureLow()));
+      }
+    } else {
+      for (DailyDataEntity dailyData : dailyDatas) {
+        dailyData.setWindSpeed(UnitConversionUtils.kmphToMph(dailyData.getWindSpeed()));
+        dailyData.setTemperatureHigh(
+            (int) UnitConversionUtils.celsiusToFahrenheit(dailyData.getTemperatureHigh()));
+        dailyData.setTemperatureLow(
+            (int) UnitConversionUtils.celsiusToFahrenheit(dailyData.getTemperatureLow()));
+      }
+    }
+  }
+
+  private void convertHourlyData(List<HourlyDataEntity> hourlyDatas) {
+    if (preferenceRepository.getTemperatureUnit() == TemperatureUnit.CELSIUS) {
+      for (HourlyDataEntity hourlyData : hourlyDatas) {
+        hourlyData.setTemperature(
+            (int) UnitConversionUtils.fahrenheitToCelsius(hourlyData.getTemperature()));
+      }
+    } else {
+      for (HourlyDataEntity hourlyData : hourlyDatas) {
+        hourlyData.setTemperature(
+            (int) UnitConversionUtils.celsiusToFahrenheit(hourlyData.getTemperature()));
+      }
+    }
+  }
+
+  private void convertCurrentTemperature(CurrentForecastEntity currentForecast) {
+    if (preferenceRepository.getTemperatureUnit() == TemperatureUnit.CELSIUS) {
+      currentForecast.setTemperature(
+          (int) Math.round(
+              UnitConversionUtils.fahrenheitToCelsius(currentForecast.getTemperature())));
+    } else {
+      currentForecast.setTemperature(
+          (int) Math.round(
+              UnitConversionUtils.celsiusToFahrenheit(currentForecast.getTemperature())));
+    }
   }
 }
